# docs/openapi.yml
openapi: 3.0.1
info:
  title: nfcwalker
  version: 1.0.0
paths:
  /api/admin/checkpoints:
    get:
      operationId: listCheckpoints
      parameters:
      - name: siteId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: orgId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listCheckpoints 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_CheckpointResponse_"
    post:
      operationId: createCheckpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckpointRequest"
        required: true
      responses:
        "200":
          description: createCheckpoint 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckpointResponse"
  /api/admin/checkpoints/{id}:
    get:
      operationId: getCheckpoint
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getCheckpoint 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckpointResponse"
    put:
      operationId: updateCheckpoint
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCheckpointRequest"
        required: true
      responses:
        "200":
          description: updateCheckpoint 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckpointResponse"
    delete:
      operationId: deleteCheckpoint
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deleteCheckpoint 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/admin/routes:
    get:
      operationId: listRoutes
      parameters:
      - name: orgId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: siteId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listRoutes 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_RouteResponse_"
    post:
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
        required: true
      responses:
        "200":
          description: createRoute 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
  /api/admin/routes/{id}:
    get:
      operationId: getRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getRoute 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
    put:
      operationId: updateRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouteRequest"
        required: true
      responses:
        "200":
          description: updateRoute 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
    delete:
      operationId: deleteRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deleteRoute 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/admin/routes/{id}/points:
    post:
      operationId: addCheckpointsToRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAddRouteCheckpointsRequest"
        required: true
      responses:
        "200":
          description: addCheckpointsToRoute 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/devices:
    get:
      operationId: getUserDevices
      parameters:
      - name: organizationId
        in: query
        required: true
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: getUserDevices 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_DeviceResponse_"
    post:
      operationId: registerDevice
      parameters:
      - name: organizationId
        in: query
        required: true
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
        required: true
      responses:
        "200":
          description: registerDevice 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
  /api/devices/{id}:
    get:
      operationId: getDevice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getDevice 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
    delete:
      operationId: revokeDevice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: revokeDevice 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
...
  /api/routes:
    get:
      operationId: listRoutes_1
      parameters:
      - name: orgId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: siteId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listRoutes_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_RouteResponse_"
  /api/routes/{id}:
    get:
      operationId: getRoute_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getRoute_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
...
components:
  schemas:
...
    AddRouteCheckpointRequest:
      required:
      - checkpointId
      - maxOffsetSec
      - minOffsetSec
      - seq
      type: object
      properties:
        checkpointId:
          type: string
          format: uuid
        seq:
          type: integer
          format: int32
        minOffsetSec:
          type: integer
          format: int32
        maxOffsetSec:
          type: integer
          format: int32
...
    BulkAddRouteCheckpointsRequest:
      required:
      - checkpoints
      type: object
      properties:
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/AddRouteCheckpointRequest"
...
    CheckpointDetailsConfig:
      required:
      - allowNotes
      - requirePhoto
      type: object
      properties:
        requirePhoto:
          type: boolean
        allowNotes:
          type: boolean
        description:
          type: string
          nullable: true
    CheckpointResponse:
      required:
      - code
      - id
      - organizationId
      - siteId
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        code:
          type: string
        geoLat:
          type: number
          nullable: true
        geoLon:
          type: number
          nullable: true
        radiusM:
          type: number
          nullable: true
        label:
          type: string
          nullable: true
        detailsConfig:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/CheckpointDetailsConfig"
        subChecks:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/SubCheckResponse"
    CreateCheckpointRequest:
      required:
      - code
      - organizationId
      - siteId
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        code:
          type: string
        geoLat:
          type: number
          nullable: true
        geoLon:
          type: number
          nullable: true
        radiusM:
          type: number
          nullable: true
        label:
          type: string
          nullable: true
        detailsConfig:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/CheckpointDetailsConfig"
        subChecks:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/SubCheckRequest"
...
    CreateRouteRequest:
      required:
      - name
      - organizationId
      - siteId
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        name:
          type: string
...
    DeviceResponse:
      required:
      - deviceId
      - id
      - registeredAt
      - status
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        metadata:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/DeviceStatus"
        registeredAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
    DeviceStatus:
      type: string
      enum:
      - ACTIVE
      - INACTIVE
      - BLOCKED
...
    Page_CheckpointResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_CheckpointResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Page_DeviceResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_DeviceResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
...
    Page_RouteResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_RouteResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
...
    Pageable:
      required:
      - mode
      - number
      - size
      - sort
      type: object
      allOf:
      - $ref: "#/components/schemas/Sort"
      - type: object
        properties:
          number:
            type: integer
            format: int32
          size:
            type: integer
            format: int32
          mode:
            $ref: "#/components/schemas/Pageable.Mode"
          sort:
            $ref: "#/components/schemas/Sort"
    Pageable.Mode:
      type: string
      enum:
      - CURSOR_NEXT
      - CURSOR_PREVIOUS
      - OFFSET
    RegisterDeviceRequest:
      required:
      - deviceId
      type: object
      properties:
        deviceId:
          type: string
        metadata:
          type: string
          nullable: true
...
    RouteResponse:
      required:
      - id
      - name
      - organizationId
      - siteId
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        name:
          type: string
...
    Slice_CheckpointResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CheckpointResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Slice_DeviceResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/DeviceResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
...
    Slice_RouteResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RouteResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
...
    Sort:
      required:
      - orderBy
      type: object
      properties:
        orderBy:
          type: array
          items:
            $ref: "#/components/schemas/Sort.Order"
    Sort.Order:
      required:
      - ascending
      - direction
      - ignoreCase
      - property
      type: object
      properties:
        ignoreCase:
          type: boolean
        direction:
          $ref: "#/components/schemas/Sort.Order.Direction"
        property:
          type: string
        ascending:
          type: boolean
    Sort.Order.Direction:
      type: string
      enum:
      - ASC
      - DESC
...
    SubCheckRequest:
      required:
      - allowNotes
      - label
      - requirePhoto
      type: object
      properties:
        label:
          type: string
        description:
          type: string
          nullable: true
        requirePhoto:
          type: boolean
        allowNotes:
          type: boolean
    SubCheckResponse:
      required:
      - allowNotes
      - id
      - label
      - requirePhoto
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        description:
          type: string
          nullable: true
        requirePhoto:
          type: boolean
        allowNotes:
          type: boolean
...
    UpdateCheckpointRequest:
      type: object
      properties:
        label:
          type: string
          nullable: true
        geoLat:
          type: number
          nullable: true
        geoLon:
          type: number
          nullable: true
        radiusM:
          type: number
          nullable: true
        detailsConfig:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/CheckpointDetailsConfig"
        subChecks:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/SubCheckRequest"
...
    UpdateRouteRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
...