# src/main/kotlin/ge/tiger8bit/controller/DeviceController.kt
package ge.tiger8bit.controller

import ge.tiger8bit.dto.DeviceResponse
import ge.tiger8bit.dto.RegisterDeviceRequest
import ge.tiger8bit.getLogger
import ge.tiger8bit.service.AccessService
import ge.tiger8bit.service.DeviceService
import io.micronaut.data.model.Page
import io.micronaut.data.model.Pageable
import io.micronaut.http.HttpResponse
import io.micronaut.http.annotation.*
import io.micronaut.security.annotation.Secured
import java.security.Principal
import java.time.format.DateTimeFormatter
import java.util.*

@Controller("/api/devices")
class DeviceController(
    private val deviceService: DeviceService,
    private val accessService: AccessService
) {
    private val logger = getLogger()
    private val formatter = DateTimeFormatter.ISO_INSTANT

    @Post
    @Secured("ROLE_WORKER")
    fun registerDevice(
        @Body request: RegisterDeviceRequest,
        @QueryValue organizationId: UUID,
        principal: Principal
    ): HttpResponse<DeviceResponse> {
        logger.info("POST /api/devices - device_id: {}, org: {}", request.deviceId, organizationId)

        val userId = UUID.fromString(principal.name)
        accessService.ensureWorkerOrBoss(userId, organizationId)

        val device = deviceService.registerDevice(
            userId = userId,
            organizationId = organizationId,
            deviceId = request.deviceId,
            metadata = request.metadata
        )
        return HttpResponse.created(device.toResponse())
    }

    @Get
    @Secured("ROLE_WORKER")
    fun getUserDevices(
        @QueryValue organizationId: UUID,
        @QueryValue(defaultValue = "0") page: Int,
        @QueryValue(defaultValue = "20") size: Int,
        principal: Principal
    ): HttpResponse<Page<DeviceResponse>> {
        logger.info("GET /api/devices - org: {}, page: {}, size: {}", organizationId, page, size)

        val userId = UUID.fromString(principal.name)
        accessService.ensureWorkerOrBoss(userId, organizationId)

        val pageable = Pageable.from(page, size)
        val devices = deviceService.getUserDevices(userId, organizationId, pageable)
        return HttpResponse.ok(devices.map { it.toResponse() })
    }

    @Get("/{id}")
    @Secured("ROLE_WORKER")
    fun getDevice(@PathVariable id: UUID, principal: Principal): HttpResponse<DeviceResponse> {
        val userId = UUID.fromString(principal.name)
        val device = deviceService.getDeviceEntity(id) ?: return HttpResponse.notFound()
        accessService.ensureWorkerOrBoss(userId, device.organizationId)
        if (device.userId != userId) {
            return HttpResponse.status(io.micronaut.http.HttpStatus.FORBIDDEN)
        }
        return HttpResponse.ok(device.toResponse())
    }

    @Delete("/{id}")
    @Secured("ROLE_WORKER")
    fun revokeDevice(id: UUID, principal: Principal): HttpResponse<Map<String, String>> {
        logger.info("DELETE /api/devices/{} - revoking", id)

        val userId = UUID.fromString(principal.name)
        val device = deviceService.getDeviceEntity(id) ?: return HttpResponse.notFound()
        accessService.ensureWorkerOrBoss(userId, device.organizationId)
        if (device.userId != userId) {
            return HttpResponse.status<Map<String, String>>(io.micronaut.http.HttpStatus.FORBIDDEN)
        }

        val success = deviceService.revokeDevice(id)
        return if (success) {
            HttpResponse.ok(mapOf("status" to "revoked"))
        } else {
            HttpResponse.notFound()
        }
    }

    private fun ge.tiger8bit.domain.Device.toResponse(): DeviceResponse {
        return DeviceResponse(
            id = this.id!!,
            deviceId = this.deviceId,
            metadata = this.metadata,
            status = this.status,
            registeredAt = this.registeredAt,
            lastUsedAt = this.lastUsedAt
        )
    }
}